- Fix: AreaUnit no longer crashes game - Change: Bridge gates now has bigger colliders - Change: Trail renderer on the swords. - Change: Swords are now the original size - Change: Health now heals 2 and has a 15% chance of spawn - Change: Everybody have very small knock back - Change: Collection radius has been slightly upped

+++ Assets/Script/Effects/Knockback.cs
@@ -7,7 +7,7 @@ public class Knockback : Effect
 	private CharacterController _cc;
 	private int _framesToKnockbackOver = 10;
 
-	public Knockback(int power) {
+	public Knockback(float power) {
 		knockbackPower = power;
 	}
